refactor: reorganize user entity
